[windows]update rtc version and ci config.

+++ windows/APIExample/install.ps1
@@ -1,7 +1,8 @@
 $ThirdPartysrc = 'https://agora-adc-artifacts.oss-cn-beijing.aliyuncs.com/libs/ThirdParty.zip'
 $ThirdPartydes = 'ThirdParty.zip'
-$agora_sdk = 'https://download.agora.io/sdk/release/Agora_Native_SDK_for_Windows_v3.7.1_FULL.zip'
+$agora_sdk = 'https://download.agora.io/sdk/release/Agora_Native_SDK_for_Windows_v3.7.2_FULL.zip'
 $agora_des = 'Agora_Native_SDK_for_Windows.zip'
+$agora_local_sdk = '../../libs'
 $MediaPlayerSDK = 'https://download.agora.io/sdk/release/Agora_Media_Player_for_Windows_x86_rel.v1.3.0_63393_ffmpeg_player_lite_20210727_1117.zip'
 $MediaPlayerSDK64 = 'https://download.agora.io/sdk/release/Agora_Media_Player_for_Windows_x64_rel.v1.3.0_63392_ffmpeg_player_lite_20210727_1117.zip'
 $MediaPlayerDes = 'MediaPlayerPartSave.zip'
@@ -20,14 +21,19 @@ if (-not (Test-Path ThirdParty)){
 
 
 if (-not (Test-Path libs)){
-	echo "download $agora_des"
-	mkdir libs
-	(New-Object System.Net.WebClient).DownloadFile($agora_sdk,$agora_des)
-	Unblock-File $agora_des
-	Expand-Archive -Path $agora_des -DestinationPath . -Force 
-	Move-Item Agora_Native_SDK_for_Windows_FULL\libs\* libs
-	Remove-Item  $agora_des -Recurse
-	Remove-Item  Agora_Native_SDK_for_Windows_FULL -Recurse
+	if (Test-Path $agora_local_sdk){
+	    	mkdir libs
+	    	Copy-Item $agora_local_sdk\*  libs -Recurse
+    	}else{
+		echo "download $agora_des"
+		mkdir libs
+		(New-Object System.Net.WebClient).DownloadFile($agora_sdk,$agora_des)
+		Unblock-File $agora_des
+		Expand-Archive -Path $agora_des -DestinationPath . -Force 
+		Move-Item Agora_Native_SDK_for_Windows_FULL\libs\* libs
+		Remove-Item  $agora_des -Recurse
+		Remove-Item  Agora_Native_SDK_for_Windows_FULL -Recurse
+	}
 }
 
 if($arch -eq 'x64'){
